@@ -1,171 +1,121 @@
 <?php
+
 namespace Config;
 
 use CodeIgniter\Config\BaseConfig;
 
 class Email extends BaseConfig
 {
+    public string $fromEmail  = '';
+    public string $fromName   = '';
+    public string $recipients = '';
 
-	/**
-	 * @var string
-	 */
-	public $fromEmail;
+    /**
+     * The "user agent"
+     */
+    public string $userAgent = 'CodeIgniter';
 
-	/**
-	 * @var string
-	 */
-	public $fromName;
+    /**
+     * The mail sending protocol: mail, sendmail, smtp
+     */
+    public string $protocol = 'mail';
 
-	/**
-	 * @var string
-	 */
-	public $recipients;
+    /**
+     * The server path to Sendmail.
+     */
+    public string $mailPath = '/usr/sbin/sendmail';
 
-	/**
-	 * The "user agent"
-	 *
-	 * @var string
-	 */
-	public $userAgent = 'CodeIgniter';
+    /**
+     * SMTP Server Hostname
+     */
+    public string $SMTPHost = '';
 
-	/**
-	 * The mail sending protocol: mail, sendmail, smtp
-	 *
-	 * @var string
-	 */
-	public $protocol = 'mail';
+    /**
+     * SMTP Username
+     */
+    public string $SMTPUser = '';
 
-	/**
-	 * The server path to Sendmail.
-	 *
-	 * @var string
-	 */
-	public $mailPath = '/usr/sbin/sendmail';
+    /**
+     * SMTP Password
+     */
+    public string $SMTPPass = '';
 
-	/**
-	 * SMTP Server Address
-	 *
-	 * @var string
-	 */
-	public $SMTPHost;
+    /**
+     * SMTP Port
+     */
+    public int $SMTPPort = 25;
 
-	/**
-	 * SMTP Username
-	 *
-	 * @var string
-	 */
-	public $SMTPUser;
+    /**
+     * SMTP Timeout (in seconds)
+     */
+    public int $SMTPTimeout = 5;
 
-	/**
-	 * SMTP Password
-	 *
-	 * @var string
-	 */
-	public $SMTPPass;
+    /**
+     * Enable persistent SMTP connections
+     */
+    public bool $SMTPKeepAlive = false;
 
-	/**
-	 * SMTP Port
-	 *
-	 * @var integer
-	 */
-	public $SMTPPort = 25;
+    /**
+     * SMTP Encryption.
+     *
+     * @var string '', 'tls' or 'ssl'. 'tls' will issue a STARTTLS command
+     *             to the server. 'ssl' means implicit SSL. Connection on port
+     *             465 should set this to ''.
+     */
+    public string $SMTPCrypto = 'tls';
 
-	/**
-	 * SMTP Timeout (in seconds)
-	 *
-	 * @var integer
-	 */
-	public $SMTPTimeout = 5;
+    /**
+     * Enable word-wrap
+     */
+    public bool $wordWrap = true;
 
-	/**
-	 * Enable persistent SMTP connections
-	 *
-	 * @var boolean
-	 */
-	public $SMTPKeepAlive = false;
+    /**
+     * Character count to wrap at
+     */
+    public int $wrapChars = 76;
 
-	/**
-	 * SMTP Encryption. Either tls or ssl
-	 *
-	 * @var string
-	 */
-	public $SMTPCrypto = 'tls';
+    /**
+     * Type of mail, either 'text' or 'html'
+     */
+    public string $mailType = 'text';
 
-	/**
-	 * Enable word-wrap
-	 *
-	 * @var boolean
-	 */
-	public $wordWrap = true;
+    /**
+     * Character set (utf-8, iso-8859-1, etc.)
+     */
+    public string $charset = 'UTF-8';
 
-	/**
-	 * Character count to wrap at
-	 *
-	 * @var integer
-	 */
-	public $wrapChars = 76;
+    /**
+     * Whether to validate the email address
+     */
+    public bool $validate = false;
 
-	/**
-	 * Type of mail, either 'text' or 'html'
-	 *
-	 * @var string
-	 */
-	public $mailType = 'text';
+    /**
+     * Email Priority. 1 = highest. 5 = lowest. 3 = normal
+     */
+    public int $priority = 3;
 
-	/**
-	 * Character set (utf-8, iso-8859-1, etc.)
-	 *
-	 * @var string
-	 */
-	public $charset = 'UTF-8';
+    /**
+     * Newline character. (Use “\r\n” to comply with RFC 822)
+     */
+    public string $CRLF = "\r\n";
 
-	/**
-	 * Whether to validate the email address
-	 *
-	 * @var boolean
-	 */
-	public $validate = false;
+    /**
+     * Newline character. (Use “\r\n” to comply with RFC 822)
+     */
+    public string $newline = "\r\n";
 
-	/**
-	 * Email Priority. 1 = highest. 5 = lowest. 3 = normal
-	 *
-	 * @var integer
-	 */
-	public $priority = 3;
+    /**
+     * Enable BCC Batch Mode.
+     */
+    public bool $BCCBatchMode = false;
 
-	/**
-	 * Newline character. (Use “\r\n” to comply with RFC 822)
-	 *
-	 * @var string
-	 */
-	public $CRLF = "\r\n";
+    /**
+     * Number of emails in each BCC batch
+     */
+    public int $BCCBatchSize = 200;
 
-	/**
-	 * Newline character. (Use “\r\n” to comply with RFC 822)
-	 *
-	 * @var string
-	 */
-	public $newline = "\r\n";
-
-	/**
-	 * Enable BCC Batch Mode.
-	 *
-	 * @var boolean
-	 */
-	public $BCCBatchMode = false;
-
-	/**
-	 * Number of emails in each BCC batch
-	 *
-	 * @var integer
-	 */
-	public $BCCBatchSize = 200;
-
-	/**
-	 * Enable notify message from server
-	 *
-	 * @var boolean
-	 */
-	public $DSN = false;
-
+    /**
+     * Enable notify message from server
+     */
+    public bool $DSN = false;
 }
