@@ -1,34 +1,42 @@
-<?php namespace Config;
+<?php
+
+namespace Config;
 
 use CodeIgniter\Config\BaseConfig;
 
 class Honeypot extends BaseConfig
 {
+    /**
+     * Makes Honeypot visible or not to human
+     */
+    public bool $hidden = true;
 
-	/**
-	 * Makes Honeypot visible or not to human
-	 *
-	 * @var boolean
-	 */
-	public $hidden = true;
-	/**
-	 * Honeypot Label Content
-	 *
-	 * @var string
-	 */
-	public $label = 'Fill This Field';
+    /**
+     * Honeypot Label Content
+     */
+    public string $label = 'Fill This Field';
 
-	/**
-	 * Honeypot Field Name
-	 *
-	 * @var string
-	 */
-	public $name = 'honeypot';
+    /**
+     * Honeypot Field Name
+     */
+    public string $name = 'honeypot';
 
-	/**
-	 * Honeypot HTML Template
-	 *
-	 * @var string
-	 */
-	public $template = '<label>{label}</label><input type="text" name="{name}" value=""/>';
+    /**
+     * Honeypot HTML Template
+     */
+    public string $template = '<label>{label}</label><input type="text" name="{name}" value="">';
+
+    /**
+     * Honeypot container
+     *
+     * If you enabled CSP, you can remove `style="display:none"`.
+     */
+    public string $container = '<div style="display:none">{template}</div>';
+
+    /**
+     * The id attribute for Honeypot container tag
+     *
+     * Used when CSP is enabled.
+     */
+    public string $containerId = 'hpc';
 }
