@@ -1,31 +1,31 @@
-<?php namespace Config;
+<?php
+
+namespace Config;
 
 use CodeIgniter\Config\BaseConfig;
+use CodeIgniter\Images\Handlers\GDHandler;
+use CodeIgniter\Images\Handlers\ImageMagickHandler;
 
 class Images extends BaseConfig
 {
-	/**
-	 * Default handler used if no other handler is specified.
-	 *
-	 * @var string
-	 */
-	public $defaultHandler = 'gd';
+    /**
+     * Default handler used if no other handler is specified.
+     */
+    public string $defaultHandler = 'gd';
 
-	/**
-	 * The path to the image library.
-	 * Required for ImageMagick, GraphicsMagick, or NetPBM.
-	 *
-	 * @var string
-	 */
-	public $libraryPath = '/usr/local/bin/convert';
+    /**
+     * The path to the image library.
+     * Required for ImageMagick, GraphicsMagick, or NetPBM.
+     */
+    public string $libraryPath = '/usr/local/bin/convert';
 
-	/**
-	 * The available handler classes.
-	 *
-	 * @var array
-	 */
-	public $handlers = [
-		'gd'      => \CodeIgniter\Images\Handlers\GDHandler::class,
-		'imagick' => \CodeIgniter\Images\Handlers\ImageMagickHandler::class,
-	];
+    /**
+     * The available handler classes.
+     *
+     * @var array<string, string>
+     */
+    public array $handlers = [
+        'gd'      => GDHandler::class,
+        'imagick' => ImageMagickHandler::class,
+    ];
 }
