@@ -1,62 +1,63 @@
-<?php namespace Config;
+<?php
 
-use CodeIgniter\Config\BaseConfig;
-use Kint\Renderer\Renderer;
+namespace Config;
 
-class Kint extends BaseConfig
+use Kint\Parser\ConstructablePluginInterface;
+use Kint\Renderer\Rich\TabPluginInterface;
+use Kint\Renderer\Rich\ValuePluginInterface;
+
+/**
+ * --------------------------------------------------------------------------
+ * Kint
+ * --------------------------------------------------------------------------
+ *
+ * We use Kint's `RichRenderer` and `CLIRenderer`. This area contains options
+ * that you can set to customize how Kint works for you.
+ *
+ * @see https://kint-php.github.io/kint/ for details on these settings.
+ */
+class Kint
 {
-	/*
-	|--------------------------------------------------------------------------
-	| Kint
-	|--------------------------------------------------------------------------
-	|
-	| We use Kint's RichRenderer and CLIRenderer. This area contains options
-	| that you can set to customize how Kint works for you.
-	|
-	| For details on these settings, see Kint's docs:
-	|	https://kint-php.github.io/kint/
-	|
-	*/
+    /*
+    |--------------------------------------------------------------------------
+    | Global Settings
+    |--------------------------------------------------------------------------
+    */
 
-	/*
-	|--------------------------------------------------------------------------
-	| Global Settings
-	|--------------------------------------------------------------------------
-	*/
+    /**
+     * @var list<class-string<ConstructablePluginInterface>|ConstructablePluginInterface>|null
+     */
+    public $plugins;
 
-	public $plugins = null;
+    public int $maxDepth           = 6;
+    public bool $displayCalledFrom = true;
+    public bool $expanded          = false;
 
-	public $maxDepth = 6;
+    /*
+    |--------------------------------------------------------------------------
+    | RichRenderer Settings
+    |--------------------------------------------------------------------------
+    */
+    public string $richTheme = 'aante-light.css';
+    public bool $richFolder  = false;
 
-	public $displayCalledFrom = true;
+    /**
+     * @var array<string, class-string<ValuePluginInterface>>|null
+     */
+    public $richObjectPlugins;
 
-	public $expanded = false;
+    /**
+     * @var array<string, class-string<TabPluginInterface>>|null
+     */
+    public $richTabPlugins;
 
-	/*
-	|--------------------------------------------------------------------------
-	| RichRenderer Settings
-	|--------------------------------------------------------------------------
-	*/
-	public $richTheme = 'aante-light.css';
-
-	public $richFolder = false;
-
-	public $richSort = Renderer::SORT_FULL;
-
-	public $richObjectPlugins = null;
-
-	public $richTabPlugins = null;
-
-	/*
-	|--------------------------------------------------------------------------
-	| CLI Settings
-	|--------------------------------------------------------------------------
-	*/
-	public $cliColors = true;
-
-	public $cliForceUTF8 = false;
-
-	public $cliDetectWidth = true;
-
-	public $cliMinWidth = 40;
+    /*
+    |--------------------------------------------------------------------------
+    | CLI Settings
+    |--------------------------------------------------------------------------
+    */
+    public bool $cliColors      = true;
+    public bool $cliForceUTF8   = false;
+    public bool $cliDetectWidth = true;
+    public int $cliMinWidth     = 40;
 }
