@@ -1,6 +1,12 @@
-<?php namespace Config;
+<?php
 
-class ForeignCharacters extends \CodeIgniter\Config\ForeignCharacters
+namespace Config;
+
+use CodeIgniter\Config\ForeignCharacters as BaseForeignCharacters;
+
+/**
+ * @immutable
+ */
+class ForeignCharacters extends BaseForeignCharacters
 {
-
 }
