@@ -1,12 +1,11 @@
-<?php namespace App\Controllers;
+<?php
+
+namespace App\Controllers;
 
 class Home extends BaseController
 {
-	public function index()
-	{
-		return view('welcome_message');
-	}
-
-	//--------------------------------------------------------------------
-
+    public function index(): string
+    {
+        return view('welcome_message');
+    }
 }
