@@ -1,42 +1,46 @@
 <?php
 
+use CodeIgniter\Test\CIUnitTestCase;
+use CodeIgniter\Test\DatabaseTestTrait;
+use Tests\Support\Database\Seeds\ExampleSeeder;
 use Tests\Support\Models\ExampleModel;
 
-class ExampleDatabaseTest extends \Tests\Support\DatabaseTestCase
+/**
+ * @internal
+ */
+final class ExampleDatabaseTest extends CIUnitTestCase
 {
-	public function setUp(): void
-	{
-		parent::setUp();
+    use DatabaseTestTrait;
 
-		// Extra code to run before each test
-	}
+    protected $seed = ExampleSeeder::class;
 
-	public function testModelFindAll()
-	{
-		$model = new ExampleModel();
+    public function testModelFindAll(): void
+    {
+        $model = new ExampleModel();
 
-		// Get every row created by ExampleSeeder
-		$objects = $model->findAll();
+        // Get every row created by ExampleSeeder
+        $objects = $model->findAll();
 
-		// Make sure the count is as expected
-		$this->assertCount(3, $objects);
-	}
+        // Make sure the count is as expected
+        $this->assertCount(3, $objects);
+    }
 
-	public function testSoftDeleteLeavesRow()
-	{
-		$model = new ExampleModel();
-		$this->setPrivateProperty($model, 'useSoftDeletes', true);
-		$this->setPrivateProperty($model, 'tempUseSoftDeletes', true);
+    public function testSoftDeleteLeavesRow(): void
+    {
+        $model = new ExampleModel();
+        $this->setPrivateProperty($model, 'useSoftDeletes', true);
+        $this->setPrivateProperty($model, 'tempUseSoftDeletes', true);
 
-		$object = $model->first();
-		$model->delete($object->id);
+        /** @var stdClass $object */
+        $object = $model->first();
+        $model->delete($object->id);
 
-		// The model should no longer find it
-		$this->assertNull($model->find($object->id));
+        // The model should no longer find it
+        $this->assertNull($model->find($object->id));
 
-		// ... but it should still be in the database
-		$result = $model->builder()->where('id', $object->id)->get()->getResult();
+        // ... but it should still be in the database
+        $result = $model->builder()->where('id', $object->id)->get()->getResult();
 
-		$this->assertCount(1, $result);
-	}
+        $this->assertCount(1, $result);
+    }
 }
