@@ -1,33 +1,49 @@
 <?php
 
-class HealthTest extends \CodeIgniter\Test\CIUnitTestCase
+use CodeIgniter\Test\CIUnitTestCase;
+use Config\App;
+use Tests\Support\Libraries\ConfigReader;
+
+/**
+ * @internal
+ */
+final class HealthTest extends CIUnitTestCase
 {
-	public function setUp(): void
-	{
-		parent::setUp();
-	}
+    public function testIsDefinedAppPath(): void
+    {
+        $this->assertTrue(defined('APPPATH'));
+    }
 
-	public function testIsDefinedAppPath()
-	{
-		$test = defined('APPPATH');
+    public function testBaseUrlHasBeenSet(): void
+    {
+        $validation = service('validation');
 
-		$this->assertTrue($test);
-	}
+        $env = false;
 
-	public function testBaseUrlHasBeenSet()
-	{
-		$env = $config = false;
+        // Check the baseURL in .env
+        if (is_file(HOMEPATH . '.env')) {
+            $env = preg_grep('/^app\.baseURL = ./', file(HOMEPATH . '.env')) !== false;
+        }
 
-		// First check in .env
-		if (is_file(HOMEPATH . '.env'))
-		{
-			$env = (bool) preg_grep("/^app\.baseURL = './", file(HOMEPATH . '.env'));
-		}
+        if ($env) {
+            // BaseURL in .env is a valid URL?
+            // phpunit.xml.dist sets app.baseURL in $_SERVER
+            // So if you set app.baseURL in .env, it takes precedence
+            $config = new App();
+            $this->assertTrue(
+                $validation->check($config->baseURL, 'valid_url'),
+                'baseURL "' . $config->baseURL . '" in .env is not valid URL',
+            );
+        }
 
-		// Then check the actual config file
-		$reader = new \Tests\Support\Libraries\ConfigReader();
-		$config = ! empty($reader->baseUrl);
+        // Get the baseURL in app/Config/App.php
+        // You can't use Config\App, because phpunit.xml.dist sets app.baseURL
+        $reader = new ConfigReader();
 
-		$this->assertTrue($env || $config);
-	}
+        // BaseURL in app/Config/App.php is a valid URL?
+        $this->assertTrue(
+            $validation->check($reader->baseURL, 'valid_url'),
+            'baseURL "' . $reader->baseURL . '" in app/Config/App.php is not valid URL',
+        );
+    }
 }
